requirements.txt
streamlit==1.22.0
tensorflow==2.10.0
numpy==1.23.5
opencv-python==4.6.0.66
matplotlib==3.6.2
seaborn==0.12.1
pillow==9.3.0
scikit-learn==1.2.0
keras-tuner==1.1.3
